Add files via upload
almost a letter to Alex Pine. he told me it is a variant of non-parametric regression. he also said that the choice of kernel is usually unimportant - funny, given that all i did was prove that a certain kernel is optimal in this specific case of correlated time series (but uncorrelated times).

dated - 01.12.2023.