Add files via upload
almost a letter to Alex Pine. he told me it is a variant of non-parametric regression. he also said that the choice of kernel is usually unimportant - funny, given that all i did was prove that a certain kernel is optimal in this specific case of correlated time series (but uncorrelated times).

i came up with this problem when trying to estimate geostrophic wind from pressure measurements (in september 2020). the usual triangulation method seemed to be very sensitive to the random noise and give ridiculously inflated values. Maksims Pogumirskis made his own new method and didn't really like mine. i have made some (much less well-written) descriptions of this in the past.

dated - 10.11.2023.